Add permissions and connect them to role
- create entity, controller, repository, dto, service and its implementation
- implement for permission same functions as for role
- add set of permissions in Role entity to connect them via new table - role_permissions
- add permissions connecting to role while in createRole() method in service
- add checking for users with exact role in before deleting it in deleteRole() method in service
- and little updates in other files

+++ src/main/java/kg/something/events_app_backend/service/impl/RoleServiceImpl.java
@@ -1,30 +1,43 @@
 package kg.something.events_app_backend.service.impl;
 
 import kg.something.events_app_backend.dto.RoleDto;
+import kg.something.events_app_backend.entity.Permission;
 import kg.something.events_app_backend.entity.Role;
+import kg.something.events_app_backend.exception.InvalidRequestException;
 import kg.something.events_app_backend.exception.ResourceAlreadyExistsException;
 import kg.something.events_app_backend.exception.ResourceNotFoundException;
 import kg.something.events_app_backend.repository.RoleRepository;
+import kg.something.events_app_backend.service.PermissionService;
 import kg.something.events_app_backend.service.RoleService;
+import kg.something.events_app_backend.service.UserService;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Set;
 import java.util.UUID;
 
 @Service
 public class RoleServiceImpl implements RoleService {
 
     private final RoleRepository repository;
+    private final PermissionService permissionService;
+    private final UserService userService;
 
-    public RoleServiceImpl(RoleRepository repository) {
+    public RoleServiceImpl(RoleRepository repository, PermissionService permissionService, @Lazy UserService userService) {
         this.repository = repository;
+        this.permissionService = permissionService;
+        this.userService = userService;
     }
 
+    @Transactional
     public String createRole(RoleDto role) {
         if (repository.existsByName(role.getName())) {
             throw new ResourceAlreadyExistsException("Роль с названием '" + role.getName() + "' уже есть в базе данных");
         }
-        repository.save(new Role(role.getName()));
+        Set<Permission> permissions = permissionService.getExistingPermissions(role.getPermissions());
+        repository.save(new Role(role.getName(), permissions));
 
         return "Роль '" + role.getName() + "' сохранена";
     }
@@ -34,6 +47,10 @@ public class RoleServiceImpl implements RoleService {
         if (role == null) {
             throw new ResourceNotFoundException("Роль с id " + id + " не найдена");
         }
+        if (!userService.findUsersByRoleId(role.getId()).isEmpty()) {
+            throw new InvalidRequestException("Удаление роли запрещено. Роль привязана к пользователям");
+        }
+        role.getPermissions().clear();
         repository.delete(role);
 
         return "Роль '" + role.getName() + "' удалена";
@@ -65,16 +82,16 @@ public class RoleServiceImpl implements RoleService {
             throw new ResourceNotFoundException("Роль с id " + id + " не найдена");
         }
 
-        String oldRoleName = role.getName();
-        if (role.getName().equals(roleDto.getName())) {
-            return "Роль не изменена. Данные из запроса и записи в базе данных идентичны";
+        if (repository.existsByName(roleDto.getName()) && !roleDto.getName().equals(role.getName())) {
+            throw new ResourceAlreadyExistsException("Роль с названием '" + roleDto.getName() + "' уже есть в базе данных");
         }
-        if (repository.existsByName(roleDto.getName())) {
-            throw new ResourceAlreadyExistsException("Роль с названием '" + role.getName() + "' уже есть в базе данных");
-        }
+        Set<Permission> permissions = permissionService.getExistingPermissions(roleDto.getPermissions());
+
         role.setName(roleDto.getName());
+        role.getPermissions().clear();
+        role.setPermissions(permissions);
         repository.save(role);
 
-        return "Название категории изменено с '" + oldRoleName + "' на '" + role.getName() + "'";
+        return "Изменения роли внесены в базу данных";
     }
 }
